Splitted accommodations controller to booking and availabilities controller
